style: controller(servlet) 및 session 패키지 구조 변경
