+++ tomcat/src/main/java/org/apache/catalina/controller/support/FileFinder.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller.support;
+package org.apache.catalina.controller.support;
 
 import java.io.File;
 import java.io.IOException;
