+++ tomcat/src/test/java/org/apache/catalina/controller/PageControllerTest.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.controller;
+package org.apache.catalina.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.HashMap;
-import org.apache.coyote.http11.controller.support.FileFinder;
+import org.apache.catalina.controller.support.FileFinder;
 import org.apache.coyote.http11.httpmessage.HttpHeader;
 import org.apache.coyote.http11.httpmessage.request.HttpMethod;
 import org.apache.coyote.http11.httpmessage.request.HttpRequest;
