+++ tomcat/src/test/java/org/apache/catalina/controller/RootControllerTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package org.apache.catalina.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
